+++ src/app/app.module.ts
@@ -13,10 +13,10 @@ import { SearchComponent } from './video-courses';
 import { CourseComponent } from './video-courses';
 import { CoursesComponent } from './video-courses';
 import { VideoCoursesComponent } from './video-courses';
-import { DateHighlightDirective } from './video-courses/courses/course/date-highlight/date-highlight.directive';
-import { DurationFormattingPipe } from './video-courses/courses/course/duration-formatting/duration-formatting.pipe';
-import { OrderByPipe } from './video-courses/courses/order-by/order-by.pipe';
-import { FilterPipe } from './video-courses/search/filter-pipe/filter.pipe';
+import { DateHighlightDirective } from './shared/directives/date-highlight/date-highlight.directive';
+import { DurationFormattingPipe } from './shared/pipes/duration-formatting/duration-formatting.pipe';
+import { OrderByPipe } from './shared/pipes/order-by/order-by.pipe';
+import { FilterPipe } from './shared/pipes/filter-pipe/filter.pipe';
 
 @NgModule({
   declarations: [
