+++ src/app/shared/directives/date-highlight/constants.ts
@@ -1,0 +1,2 @@
+
+export const twoWeeksMs: number = 1.2096e9;
