+++ src/app/shared/pipes/duration-formatting/duration-formatting.pipe.ts
@@ -9,7 +9,7 @@ export class DurationFormattingPipe implements PipeTransform {
     const hours: number = value / 60;
     if (hours >= 1) {
       const rest: number = value % 60;
-      return `${Math.trunc(hours)}h` + (rest > 0 ? ` ${rest}min` : '');
+      return `${Math.trunc(hours)}h${rest > 0 ? ` ${rest}min` : ''}`;
     }
     return `${value}min`;
   }
